EcoNewsTagFilterComponent Refactoring,TagButton enum added
