+++ src/main/java/com/greencity/ui/component/ecoNewsTag/TagButton.java
@@ -1,0 +1,10 @@
+package com.greencity.ui.component.ecoNewsTag;
+
+public enum TagButton {
+    NEWS(),
+    EVENTS(),
+    EDUCATION(),
+    INITIATIVES(),
+    ADS()
+}
+
